thrift proxy: move UpstreamRequest into its own file
This is needed so that it can be reused by the upcoming shadowing
implementation, which will introduce a dependency from Router to
ShadowRouter to UpstreamRequest, so this breaks the circular
dependency.

I have one more PR (4/5) that moves common functionality from
UpstreamRequest into RequestOwner. After that I'll send over
the actual shadow routing implementation, which reuses all of
these parts.
