+++ source/extensions/filters/network/thrift_proxy/router/BUILD
@@ -46,12 +46,28 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "upstream_request_lib",
+    srcs = ["upstream_request.cc"],
+    hdrs = ["upstream_request.h"],
+    deps = [
+        ":router_interface",
+        "//envoy/tcp:conn_pool_interface",
+        "//source/extensions/filters/network/thrift_proxy:app_exception_lib",
+        "//source/extensions/filters/network/thrift_proxy:conn_manager_lib",
+        "//source/extensions/filters/network/thrift_proxy:thrift_object_interface",
+        "//source/extensions/filters/network/thrift_proxy:transport_interface",
+        "//source/extensions/filters/network/thrift_proxy/filters:filter_interface",
+    ],
+)
+
+envoy_cc_library(
     name = "router_lib",
     srcs = ["router_impl.cc"],
     hdrs = ["router_impl.h"],
     deps = [
         ":router_interface",
         ":router_ratelimit_lib",
+        ":upstream_request_lib",
         "//envoy/tcp:conn_pool_interface",
         "//envoy/upstream:cluster_manager_interface",
         "//envoy/upstream:load_balancer_interface",
